add best model selector
